+++ framework/src/main/java/org/checkerframework/framework/source/SourceChecker.java
@@ -430,7 +430,7 @@ public abstract class SourceChecker extends AbstractTypeProcessor implements Opt
 
   /**
    * Maps error keys to localized/custom error messages. Do not use directly; call {@link
-   * #fullMessageOf} or {@link #processArg}. Is set in {@link #initChecker}.
+   * #fullMessageOf} or {@link #processErrorMessageArg}. Is set in {@link #initChecker}.
    */
   protected Properties messagesProperties;
 
@@ -1118,7 +1118,7 @@ public abstract class SourceChecker extends AbstractTypeProcessor implements Opt
 
     if (args != null) {
       for (int i = 0; i < args.length; ++i) {
-        args[i] = processArg(args[i]);
+        args[i] = processErrorMessageArg(args[i]);
       }
     }
 
@@ -1311,7 +1311,7 @@ public abstract class SourceChecker extends AbstractTypeProcessor implements Opt
    * @param arg the argument
    * @return the result after processing
    */
-  protected Object processArg(Object arg) {
+  protected Object processErrorMessageArg(Object arg) {
     // Check to see if the argument itself is a property to be expanded
     if (arg instanceof String) {
       return messagesProperties.getProperty((String) arg, (String) arg);
@@ -2082,8 +2082,8 @@ public abstract class SourceChecker extends AbstractTypeProcessor implements Opt
   }
 
   /**
-   * Determines whether all the warnings pertaining to a given tree should be suppressed. Returns
-   * true if the tree is within the scope of a @SuppressWarnings annotation, one of whose values
+   * Returns true if all the warnings pertaining to a given tree should be suppressed. Returns true
+   * if the tree is within the scope of a @SuppressWarnings annotation, one of whose values
    * suppresses the checker's warning. Also, returns true if the {@code errKey} matches a string in
    * {@code -AsuppressWarnings}.
    *
@@ -2115,9 +2115,9 @@ public abstract class SourceChecker extends AbstractTypeProcessor implements Opt
   }
 
   /**
-   * Determines whether all the warnings pertaining to a given tree path should be suppressed.
-   * Returns true if the path is within the scope of a @SuppressWarnings annotation, one of whose
-   * values suppresses the checker's warning.
+   * Returns true if all the warnings pertaining to a given tree path should be suppressed. Returns
+   * true if the path is within the scope of a @SuppressWarnings annotation, one of whose values
+   * suppresses the checker's warning.
    *
    * @param path the TreePath that might be a source of, or related to, a warning
    * @param errKey the error key the checker is emitting
@@ -2211,7 +2211,7 @@ public abstract class SourceChecker extends AbstractTypeProcessor implements Opt
   protected final Set<Element> elementsWithSuppressedWarnings = new HashSet<>();
 
   /**
-   * Determines whether all the warnings pertaining to a given element should be suppressed. Returns
+   * Returns true if all the warnings pertaining to a given element should be suppressed. Returns
    * true if the element is within the scope of a @SuppressWarnings annotation, one of whose values
    * suppresses all the checker's warnings.
    *
@@ -2248,8 +2248,8 @@ public abstract class SourceChecker extends AbstractTypeProcessor implements Opt
   }
 
   /**
-   * Determines whether an error (whose message key is {@code messageKey}) should be suppressed. It
-   * is suppressed if any of the given SuppressWarnings strings suppresses it.
+   * Returns true if an error (whose message key is {@code messageKey}) should be suppressed. It is
+   * suppressed if any of the given SuppressWarnings strings suppresses it.
    *
    * <p>A SuppressWarnings string may be of the following pattern:
    *