+++ framework/src/main/java/org/checkerframework/framework/type/AnnotatedTypeFactory.java
@@ -318,14 +318,14 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
    * Caches the supported type qualifier classes. Call {@link #getSupportedTypeQualifiers()} instead
    * of using this field directly, as it may not have been initialized.
    */
-  private final Set<Class<? extends Annotation>> supportedQuals;
+  private @MonotonicNonNull Set<Class<? extends Annotation>> supportedQuals = null;
 
   /**
    * Caches the fully-qualified names of the classes in {@link #supportedQuals}. Call {@link
    * #getSupportedTypeQualifierNames()} instead of using this field directly, as it may not have
    * been initialized.
    */
-  private final Set<@CanonicalName String> supportedQualNames;
+  private @MonotonicNonNull Set<@CanonicalName String> supportedQualNames = null;
 
   /** Parses stub files and stores annotations on public elements from stub files. */
   public final AnnotationFileElementTypes stubTypes;
@@ -563,8 +563,6 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
     this.elements = processingEnv.getElementUtils();
     this.types = processingEnv.getTypeUtils();
 
-    this.supportedQuals = new HashSet<>();
-    this.supportedQualNames = new HashSet<>();
     this.stubTypes = new AnnotationFileElementTypes(this);
     this.ajavaTypes = new AnnotationFileElementTypes(this);
     this.currentFileAjavaTypes = null;
@@ -705,14 +703,14 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
   }
 
   /**
-   * Requires that supportedQuals is non-empty and each element is a type qualifier. That is, no
-   * element has a {@code @Target} meta-annotation that contains something besides TYPE_USE or
-   * TYPE_PARAMETER. (@Target({}) is allowed.) @
+   * Requires that supportedQuals is non-null and non-empty and each element is a type qualifier.
+   * That is, no element has a {@code @Target} meta-annotation that contains something besides
+   * TYPE_USE or TYPE_PARAMETER. (@Target({}) is allowed.) @
    *
    * @throws BugInCF If supportedQuals is empty or contaions a non-type qualifier
    */
   private void checkSupportedQualsAreTypeQuals() {
-    if (supportedQuals.isEmpty()) {
+    if (supportedQuals == null || supportedQuals.isEmpty()) {
       throw new TypeSystemError("Found no supported qualifiers.");
     }
     for (Class<? extends Annotation> annotationClass : supportedQuals) {
@@ -1240,11 +1238,11 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
    *     supported
    */
   public final Set<Class<? extends Annotation>> getSupportedTypeQualifiers() {
-    if (this.supportedQuals.isEmpty()) {
-      supportedQuals.addAll(createSupportedTypeQualifiers());
+    if (this.supportedQuals == null) {
+      supportedQuals = createSupportedTypeQualifiers();
       checkSupportedQualsAreTypeQuals();
     }
-    return Collections.unmodifiableSet(supportedQuals);
+    return supportedQuals;
   }
 
   /**
@@ -1259,12 +1257,14 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
    *     supported
    */
   public final Set<@CanonicalName String> getSupportedTypeQualifierNames() {
-    if (this.supportedQualNames.isEmpty()) {
+    if (this.supportedQualNames == null) {
+      supportedQualNames = new HashSet<>();
       for (Class<?> clazz : getSupportedTypeQualifiers()) {
         supportedQualNames.add(clazz.getCanonicalName());
       }
+      supportedQualNames = Collections.unmodifiableSet(supportedQualNames);
     }
-    return Collections.unmodifiableSet(supportedQualNames);
+    return supportedQualNames;
   }
 
   // **********************************************************************
@@ -2780,6 +2780,8 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
     }
 
     con = (AnnotatedExecutableType) typeVarSubstitutor.substitute(typeParamToTypeArg, con);
+
+    stubTypes.injectRecordComponentType(types, ctor, con);
     if (enclosingType != null) {
       // Reset the enclosing type because it can be substituted incorrectly.
       ((AnnotatedDeclaredType) con.getReturnType()).setEnclosingType(enclosingType);
@@ -2787,8 +2789,21 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
     if (type.isUnderlyingTypeRaw()) {
       ((AnnotatedDeclaredType) con.getReturnType()).setIsUnderlyingTypeRaw();
     }
-    stubTypes.injectRecordComponentType(types, ctor, con);
+    if (ctor.getEnclosingElement().getKind() == ElementKind.ENUM) {
+      Set<AnnotationMirror> enumAnnos = getEnumConstructorQualifiers();
+      con.getReturnType().replaceAnnotations(enumAnnos);
+    }
     return new ParameterizedExecutableType(con, typeargs);
+  }
+
+  /**
+   * Returns the annotations that should be applied to enum constructors. This implementation
+   * returns an empty set. Subclasses can override to return a different set.
+   *
+   * @return the annotations that should be applied to enum constructors
+   */
+  protected Set<AnnotationMirror> getEnumConstructorQualifiers() {
+    return Collections.emptySet();
   }
 
   /**
@@ -3213,9 +3228,9 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
   // **********************************************************************
 
   /**
-   * Determines whether the given annotation is a part of the type system under which this type
-   * factory operates. Null is never a supported qualifier; the parameter is nullable to allow the
-   * result of canonicalAnnotation to be passed in directly.
+   * Returns true if the given annotation is a part of the type system under which this type factory
+   * operates. Null is never a supported qualifier; the parameter is nullable to allow the result of
+   * canonicalAnnotation to be passed in directly.
    *
    * @param a any annotation
    * @return true if that annotation is part of the type system under which this type factory
@@ -3230,7 +3245,7 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
   }
 
   /**
-   * Determines whether the given class is a part of the type system under which this type factory
+   * Returns true if the given class is a part of the type system under which this type factory
    * operates.
    *
    * @param clazz annotation class
@@ -3242,8 +3257,8 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
   }
 
   /**
-   * Determines whether the given class name is a part of the type system under which this type
-   * factory operates.
+   * Returns true if the given class name is a part of the type system under which this type factory
+   * operates.
    *
    * @param className fully-qualified annotation class name
    * @return true if that class name is a type qualifier in the type system under which this type